Test sessions list endpoint
